# tool: performance improvements