--- a/tool/microkit/src/sel4.rs
+++ b/tool/microkit/src/sel4.rs
@@ -29,7 +29,6 @@ pub struct BootInfo {
 /// initial task.
 #[derive(Clone)]
 pub struct Object {
-    pub name: String,
     /// Type of kernel object
     pub object_type: ObjectType,
     pub cap_addr: u64,