This PR implements a number of performance improvements to make sure that building a Microkit system feels instantaneous, even for larger systems.

Most of the changes decrease the two main bottlenecks, ELF parsing and dynamic memory allocation (aka `.clone` in Rust).

This PR significantly decreases the performance and memory usage of the tool.

Below is performance numbers for a hello world program (top is after, bottom is before):
```
Benchmark 1 (1530 runs): /home/ivanv/ts/microkit/release/microkit-sdk-1.3.0/bin/microkit example/qemu_virt_aarch64/hello/hello.system --search-path /home/ivanv/ts/microkit/tmp_build --board qemu_virt_aarch64 --config debug -o /home/ivanv/ts/microkit/tmp_build/loader.img -r /home/ivanv/ts/microkit/tmp_build/report.txt
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          3.23ms ±  227us    2.94ms … 10.4ms         62 ( 4%)        0%
  peak_rss           2.62MB ± 31.3KB    2.46MB … 2.62MB         58 ( 4%)        0%
  cpu_cycles         3.15M  ±  136K      277K  … 3.98M          49 ( 3%)        0%
  instructions       3.93M  ±  159K      241K  … 3.94M          32 ( 2%)        0%
  cache_references   18.5K  ± 1.40K     2.36K  … 27.6K         167 (11%)        0%
  cache_misses       1.89K  ± 1.23K       46   … 9.64K           2 ( 0%)        0%
  branch_misses      25.8K  ± 1.04K     2.76K  … 28.6K          85 ( 6%)        0%
Benchmark 2 (340 runs): /tmp/microkit/release/microkit-sdk-1.3.0/bin/microkit example/qemu_virt_aarch64/hello/hello.system --search-path /tmp/microkit/tmp_build --board qemu_virt_aarch64 --config debug -o /tmp/microkit/tmp_build/loader.img -r /tmp/microkit/tmp_build/report.txt
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          14.7ms ±  785us    13.2ms … 16.6ms          0 ( 0%)        +355.4% ±  1.4%
  peak_rss           7.18MB ± 73.6KB    6.88MB … 7.37MB         64 (19%)        +174.4% ±  0.2%
  cpu_cycles         55.4M  ± 3.83M     26.5M  … 64.1M           2 ( 1%)        +1658.4% ±  6.1%
  instructions        164M  ± 5.08M     77.1M  …  165M           9 ( 3%)        +4079.0% ±  6.5%
  cache_references   58.5K  ± 2.89K     41.4K  … 75.3K          26 ( 8%)        +216.9% ±  1.1%
  cache_misses       11.2K  ± 3.04K     5.34K  … 30.2K           8 ( 2%)        +495.0% ± 10.7%
  branch_misses      30.6K  ±  654      22.3K  … 33.3K          17 ( 5%)        + 18.6% ±  0.4%
```

Below is performance numbers aiming to represent how long it takes to build a larger system that allocates 256MB of 4K pages and has an ELF size of 80MB:
```
Benchmark 1 (16 runs): /home/ivanv/ts/microkit/release/microkit-sdk-1.3.0/bin/microkit example/qemu_virt_aarch64/hello/hello.system --search-path /home/ivanv/ts/microkit/tmp_build --board qemu_virt_aarch64 --config debug -o /home/ivanv/ts/microkit/tmp_build/loader.img -r /home/ivanv/ts/microkit/tmp_build/report.txt
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           314ms ± 13.0ms     305ms …  346ms          0 ( 0%)        0%
  peak_rss           37.8MB ±  105KB    37.7MB … 38.0MB          0 ( 0%)        0%
  cpu_cycles         1.17G  ± 4.36M     1.16G  … 1.18G           0 ( 0%)        0%
  instructions       1.37G  ± 3.84M     1.35G  … 1.37G           3 (19%)        0%
  cache_references   3.36M  ± 66.0K     3.18M  … 3.45M           1 ( 6%)        0%
  cache_misses        789K  ± 25.2K      752K  …  836K           0 ( 0%)        0%
  branch_misses       499K  ± 24.1K      476K  …  556K           2 (13%)        0%
Benchmark 2 (5 runs): /tmp/microkit/release/microkit-sdk-1.3.0/bin/microkit example/qemu_virt_aarch64/hello/hello.system --search-path /tmp/microkit/tmp_build --board qemu_virt_aarch64 --config debug -o /tmp/microkit/tmp_build/loader.img -r /tmp/microkit/tmp_build/report.txt
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          1.24s  ± 59.0ms    1.19s  … 1.34s           0 ( 0%)        +293.4% ± 10.0%
  peak_rss            264MB ±  202KB     264MB …  264MB          0 ( 0%)        +597.7% ±  0.4%
  cpu_cycles         5.41G  ±  253M     5.18G  … 5.85G           0 ( 0%)        +364.0% ± 10.7%
  instructions       17.1G  ±  662K     17.1G  … 17.1G           0 ( 0%)        +1149.1% ±  0.3%
  cache_references   13.4M  ± 60.5K     13.3M  … 13.4M           0 ( 0%)        +297.0% ±  2.1%
  cache_misses       2.88M  ± 83.6K     2.75M  … 2.96M           0 ( 0%)        +265.0% ±  6.0%
  branch_misses       706K  ± 90.7K      655K  …  866K           0 ( 0%)        + 41.5% ± 10.1%
```